feats_v2-geographic-diversity-of-india - initail commit with confer project files changes for generic adjustment
