+++ src/confer/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import { AppProvider } from './social_customs/AppContext';
-import AdminPage from './social_customs/AdminPage';
+import { AppProvider } from './AppContext';
+import AdminPage from './admin/AdminPage';
 import AudiencePage from './audience/AudiencePage';
 
 function App() {
@@ -9,7 +9,7 @@ function App() {
         <AppProvider>
             <Router>
                 <Routes>
-                    <Route path="confer/socialcustoms" element={<AdminPage />} />
+                    <Route path="confer/admin" element={<AdminPage />} />
                     <Route path="confer/audience" element={<AudiencePage />} />
                 </Routes>
             </Router>
